chore: move layout folder
