+++ src/lib/components/auth/AuthPopover.tsx
@@ -22,7 +22,7 @@ import { BsPencil } from "react-icons/bs";
 import { ImCheckmark } from "react-icons/im";
 import { IoMdPerson } from "react-icons/io";
 
-import { PRIVATE_ROUTES } from "lib/components/layout/RouteWrapper";
+import { PRIVATE_ROUTES } from "lib/layout/RouteWrapper";
 import SpokerInput from "lib/components/shared/SpokerInput";
 import { logoutUser, updateDisplayName } from "lib/services/firebase";
 import { roomsData } from "lib/services/firebase/room";
