+++ src/lib/components/auth/AuthWrapper.tsx
@@ -10,7 +10,7 @@ import { useRouter } from "next/router";
 import type { ReactNode } from "react";
 import { useContext, useEffect, useState } from "react";
 
-import { PUBLIC_ROUTES } from "lib/components/layout/RouteWrapper";
+import { PUBLIC_ROUTES } from "lib/layout/RouteWrapper";
 
 import { AuthContext } from "./AuthProvider";
 import Login from "./Login";
